Adds Divine RPG for DeadSpaceman Minecraft
Divine RPG is a mod that adds in new mobs, new dimensions, a great deal of new armor and weapons, and new blocks as well. While the modpack itself won't provide much in the way of additions to people's machine set-ups, the mod will add in encounters, creatures, and bosses for people to engage with, an aspect of the modpack that is so far lacking.

As for the ores and various materials that can be harvested from the modpack (most of which are in the overworld anyhow), some of the new mobs will drop them as well, helping to reduce the shortage of material.

Watch out for pumpkin spiders.